Cambios en la de oxigenación

+++ src/blood-oxygenation/application/CreateBloodOxygenation_UseCase copy.go
@@ -13,7 +13,7 @@ func NewCreateBloodOxygenation(db domain.IBloodOxygenation, r domain.IBloodOxyge
 	return &CreateBloodOxygenation{db: db, rabbit: r}
 }
 
-func (cu *CreateBloodOxygenation) Execute(esp32ID string, tiempo string, ir int32, red int32, spo2 float32) error {
+func (cu *CreateBloodOxygenation) Execute(esp32ID string, tiempo string, ir int32, red int32, spo2 float64) error {
 	err := cu.db.SaveBloodOxygenation(esp32ID, tiempo, ir, red, spo2)
 	if err != nil {
 		return err
